pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brain-forge"
version = "0.1.0-dev"
description = "Advanced Multi-Modal Brain-Computer Interface System"
readme = "README.md"
license = {text = "TBD"}
authors = [
    {name = "Brain-Forge Development Team", email = "dev@brain-forge.org"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Human Machine Interfaces",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent"
]
requires-python = ">=3.8"
dependencies = [
    # Core scientific computing
    "numpy>=1.21.0",
    "scipy>=1.7.0", 
    "scikit-learn>=1.0.0",
    "pandas>=1.3.0",
    
    # Signal processing
    "PyWavelets>=1.3.0",
    "mne>=1.0.0",
    "nilearn>=0.9.0",
    
    # Real-time streaming
    "pylsl>=1.16.0",
    
    # Neural simulation
    "brian2>=2.5.0",
    "nest-simulator>=3.0",
    
    # Configuration and logging
    "pyyaml>=6.0",
    "pydantic>=1.9.0",
    
    # Utilities
    "tqdm>=4.64.0",
    "click>=8.0.0",
    "rich>=12.0.0"
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pytest-asyncio>=0.20.0",
    "black>=22.0.0",
    "isort>=5.10.0", 
    "flake8>=4.0.0",
    "mypy>=0.950",
    "pre-commit>=2.17.0"
]

# GPU acceleration
gpu = [
    "cupy-cuda11x>=10.0.0",
    "torch>=1.11.0",
    "tensorflow>=2.8.0"
]

# Visualization
viz = [
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "plotly>=5.6.0",
    "pyvista>=0.34.0",
    "bokeh>=2.4.0"
]

# Hardware interfaces
hardware = [
    "pyserial>=3.5",
    "bleak>=0.14.0",  # Bluetooth LE
    "keyboard>=0.13.0"
]

# Web interface
web = [
    "fastapi>=0.75.0",
    "uvicorn>=0.17.0",
    "websockets>=10.2",
    "jinja2>=3.0.0"
]

# Documentation
docs = [
    "sphinx>=4.4.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.17.0"
]

# Complete installation
all = [
    "brain-forge[dev,gpu,viz,hardware,web,docs]"
]

[project.urls]
Homepage = "https://github.com/brain-forge/brain-forge"
Repository = "https://github.com/brain-forge/brain-forge.git"
Documentation = "https://brain-forge.readthedocs.io/"
"Bug Tracker" = "https://github.com/brain-forge/brain-forge/issues"

[project.scripts]
brain-forge = "src.api.cli:main"
bf-stream = "src.acquisition.stream_manager:cli_main"
bf-process = "src.processing:cli_main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--cov=src",
    "--cov-report=html",
    "--cov-report=term-missing"
]
markers = [
    "slow: marks tests as slow",
    "hardware: marks tests requiring hardware",
    "integration: marks tests as integration tests",
    "gpu: marks tests requiring GPU support"
]
